progress-events/tests/submissions/Samsung/security-consideration.sub.html

<!doctype html>
<html>
  <head>
    <title>ProgressEvent: security consideration</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/progress/raw-file/tip/Overview.html#security-considerations" data-tested-assertations="/following-sibling::p" />
  </head>
  <body>
    <div id="log"></div>
    <script>
      async_test(function() {
        var xhr = new XMLHttpRequest();

        xhr.onprogress = this.unreached_func("MUST NOT dispatch progress event.");
        xhr.onload = this.unreached_func("MUST NOT dispatch load event.");
        xhr.onerror = this.step_func(function(pe) {
          assert_equals(pe.type, "error");
          assert_equals(pe.loaded, 0, "loaded is zero.");
          assert_false(pe.lengthComputable, "lengthComputable is false.");
          assert_equals(pe.total, 0, "total is zero.");
        });
        xhr.onloadend = this.step_func_done();
        xhr.open("GET", "http://{{host}}:{{ports[http][1]}}/progress-events/tests/submissions/Samsung/resources/img.jpg", true);
        xhr.send(null);
      })
    </script>
  </body>
</html>
